src/backend/requirements.txt
flask==1.1.2
Flask-WTF==0.14.3
Jinja2==2.11.2
google-cloud-core==1.3
google-cloud-storage==1.29.0
google-cloud-firestore==1.8.1
google-cloud-pubsub==1.7
google-cloud-trace==0.23
google-python-cloud-debugger==2.15
firebase-admin==4.3.0
gunicorn==20.0.4
mock==4.0.2
opencensus==0.1.10
pytest==5.4.3
werkzeug==0.16.1
wheel==0.34.2